Extract layer, particle, and visualization modules
